src/requirements.txt
certifi==2022.5.18.1
click==8.1.3
dataclasses_json==0.5.7
Flask==2.1.2
Flask-Cors==3.0.10
Flask-SQLAlchemy==2.5.1
humanfriendly==10.0
importlib-metadata==4.11.4
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.1
six==1.16.0
Werkzeug==2.1.2
zipp==3.8.0
